% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staar_trio.R
\name{staar_trio}
\alias{staar_trio}
\title{STAAR.Trio procedure using omnibus test on basis of a family-based design}
\usage{
staar_trio(
  genotype,
  maf,
  pos,
  annotation_phred = NULL,
  rare_maf_cutoff = 0.05,
  rv_num_cutoff = 2,
  y_res = NA,
  xchr = FALSE,
  sex = NA
)
}
\arguments{
\item{genotype}{A 3n*p matrix for the original genotype data, in which n is the number of trios and p is the number of variants.
Each trio must consist of father, mother, and offspring (in this order). The genotypes must be coded as 0, 1, or 2. Missing genotypes are not allowed.}

\item{maf}{minor allele frequency}

\item{pos}{A numeric vector of length p for the position of p variants.}

\item{annotation_phred}{a data frame or matrix of functional annotation data
of dimension p*q (or a vector of a single annotation score with length p).}

\item{rare_maf_cutoff}{the cutoff of maximum minor allele frequency in
defining rare variants (default = 0.01).}

\item{rv_num_cutoff}{the cutoff of minimum number of variants of analyzing
a given variant-set (default = 2).}

\item{y_res}{A numeric vector of length n for the residual Y-Y_hat. Y_hat is the predicted value from the regression model in which the quantitative trait Y is regressed on the covariates.
If Y is dichotomous, you may treat Y as quantitative when applying the regression model.}

\item{xchr}{A logical value indicating whether the analysis is for the X chromosome. When xchr is TRUE, the analysis is for the X chromosome and sex is required.
When xchr is FALSE, the analysis is for the autosomes.}

\item{sex}{A numeric vector of length n for the sex of offspring. 0s indicate females and 1s indicate males. Sex is required when xchr is TRUE.}
}
\value{
A list for the analysis results.
}
\description{
STAAR.Trio procedure using omnibus test on basis of a family-based design
}
